requirements.txt
scapy>=2.5.0
flask>=3.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
